requirements.txt
numpy
pillow
matplotlib
jupyterlab
notebook
paddlepaddle==2.6.1
paddleocr>=2.7.0

fastapi
uvicorn[standard]
transformers>=4.43.0
torch
accelerate
safetensors
qwen-vl-utils
requests
torchvision
